LayoutTests/platform/wpe/imported/w3c/web-platform-tests/css/css-properties-values-api/typedom-expected.txt

CONSOLE MESSAGE: TypeError: undefined is not an object (evaluating 'target.attributeStyleMap.clear')

Harness Error (FAIL), message = TypeError: undefined is not an object (evaluating 'target.attributeStyleMap.clear')

FAIL Computed * is reified as CSSUnparsedValue target.computedStyleMap is not a function. (In 'target.computedStyleMap()', 'target.computedStyleMap' is undefined)
